+++ src/icrar/leap-accelerate/model/cuda/DeviceIntegration.h
@@ -65,8 +65,7 @@ namespace cuda
     {
         int m_integrationNumber;
         device_tensor3<std::complex<double>> m_visibilities; //[polarizations][baselines][channels]
-
-        size_t m_rows;
+        int64_t m_rows;
         
     public:
         /**
@@ -99,8 +98,8 @@ namespace cuda
 
         int GetIntegrationNumber() const { return m_integrationNumber; }
 
-        int GetRows() const { return m_rows; }
-        int GetChannels() const { return m_visibilities.GetDimensionSize(2); }
+        int64_t GetRows() const { return m_rows; }
+        uint64_t GetChannels() const { return m_visibilities.GetDimensionSize(2); }
         
         const device_tensor3<std::complex<double>>& GetVis() const { return m_visibilities; }
         device_tensor3<std::complex<double>>& GetVis() { return m_visibilities; }
